Piece heirarchy. Movable pieces.
Instead of enumerating piece types I decided to create a heirarchy.
There's now a rudimentary piece moving system in place. There's a visual
glitch with selecting boardlocations but I'm fairly sure I know exactly
what's wrong.

Instead of defining the togglebutton control template in a style the
template is defined as it's own key. The datatemplate of boardlocation is
now defined in ChessBoardView.xaml.