Some rays, ChessBoard model, Change toString
Rook, Knight and King now have rays. I decided to create a ChessBoard
model and abstract away behind the ChessBoardViewModel (which is what I
should have done from the beginning, I think). I changed the toString
method of the different chess pieces because the colour of the text
indicated the pieces colour.