It lives!
Wrote an engine. Currently it only works when the MAX_PLY is an odd number. I know why this is happening. The game plays against me though.

Also, one problem is that when more than one move results in the same evaluation then a random one of the moves should be chosen. Not the first.

Holy shit, I'm a little overwhelmed right now.

Even with an odd ply I know that this won't perform well with endgame because I have coded up stale or checkmate into the evaluator or engine.